pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Oleg <chaiko.oleg89@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^7.0.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
